+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/storages/energy/EnergyStorage.kt
@@ -5,9 +5,9 @@ import java.util.function.Predicate
 interface EnergyStorage : EnergySink {
     val energy: EnergyStack
     val capacity: Long
-    fun takeEnergy(predicate: Predicate<EnergyStack>, limit: Int): EnergyStack
+    fun takeEnergy(predicate: Predicate<EnergyStack>, limit: Long): EnergyStack
 
-    fun moveTo(to: EnergySink, limit: Int, takePredicate: Predicate<EnergyStack>): Long {
+    fun moveTo(to: EnergySink, limit: Long, takePredicate: Predicate<EnergyStack>): Long {
         if (movableType != null) {
             throw IllegalStateException("With movable type you should redefine this function")
         }
